Droid/requirements.txt
numpy>=2.0.0
opencv-python>=4.8.0
pyaudio>=0.2.13
SpeechRecognition>=3.10.0
pyttsx3>=2.90
requests>=2.31.0
mediapipe>=0.10.0
ollama>=0.5.0
pyserial>=3.5
